[eth]: Refactor 'tw_evm/abi' directory structure
